requirements.txt
streamlit==1.33.0
tensorflow-cpu
Pillow==8.2.0
numpy==1.19.5
